refactor: Change package name from routine to daily
- Changed package name and imports from routine to daily.
- Updated `GetDailyRoutineUseCase` and `ToggleDailyRoutineStatusUseCase`'s package.
- Modified the usage of these use cases in `HomeViewModel`.
