+++ core/domain/src/main/java/com/teampatch/core/domain/usecase/daily/ToggleDailyRoutineStatusUseCase.kt
@@ -1,4 +1,4 @@
-package com.teampatch.core.domain.usecase.routine
+package com.teampatch.core.domain.usecase.daily
 
 import javax.inject.Inject
