+++ feature/home/src/main/java/com/teampatch/feature/home/HomeViewModel.kt
@@ -15,8 +15,8 @@ import com.teampatch.core.domain.model.Todo
 import com.teampatch.core.domain.model.User
 import com.teampatch.core.domain.usecase.memory.AddMemoryCardUseCase
 import com.teampatch.core.domain.usecase.memory.GetLatestMemoryCardUseCase
-import com.teampatch.core.domain.usecase.routine.GetDailyRoutineUseCase
-import com.teampatch.core.domain.usecase.routine.ToggleDailyRoutineStatusUseCase
+import com.teampatch.core.domain.usecase.daily.GetDailyRoutineUseCase
+import com.teampatch.core.domain.usecase.daily.ToggleDailyRoutineStatusUseCase
 import com.teampatch.core.domain.usecase.user.GetUserInfoUseCase
 import com.teampatch.feature.home.model.HomeErrorHandler
 import com.teampatch.feature.home.model.MemoryCardUiState
